Rename xtensa optimized processor specific assembler code and now uses appropriate macro
